+++ pkg/controller/mysqlbackupcron/mysqlbackupcron_controller_test.go
@@ -32,7 +32,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
@@ -170,9 +169,8 @@ var _ = Describe("MysqlBackupCron controller", func() {
 
 		It("should be just one entry for a cluster", func() {
 			// update cluster spec
-			cluster.Spec.MysqlConf = map[string]intstr.IntOrString{
-				"something": intstr.FromString("new"),
-			}
+			cluster.Spec.MysqlConf = "something = new"
+
 			Expect(c.Update(context.TODO(), cluster)).To(Succeed())
 
 			// expect an reconcile event